chore: Create shared public `BaseQuestion`
