+++ apps/editor.planx.uk/src/@planx/components/Question/Public.test.tsx
@@ -1,3 +1,4 @@
+import { QuestionLayout } from "@planx/components/shared/BaseQuestion/model";
 import { act, waitFor } from "@testing-library/react";
 import { Store, useStore } from "pages/FlowEditor/lib/store";
 import React from "react";
@@ -6,7 +7,7 @@ import { vi } from "vitest";
 import { axe } from "vitest-axe";
 
 import { Option } from "../Option/model";
-import QuestionComponent, { QuestionLayout } from "./Public";
+import QuestionComponent from "./Public";
 
 const { setState } = useStore;
 
@@ -46,13 +47,13 @@ const options: { [key in QuestionLayout]: Option[] } = {
       id: "pizza_id",
       data: {
         text: "Pizza",
-      }
+      },
     },
     {
       id: "celery_id",
       data: {
         text: "Celery",
-      }
+      },
     },
   ],
   [QuestionLayout.Images]: [
@@ -61,14 +62,14 @@ const options: { [key in QuestionLayout]: Option[] } = {
       data: {
         text: "Pizza",
         img: "pizza.jpg",
-      }
+      },
     },
     {
       id: "celery_id",
       data: {
         text: "Celery",
         img: "celery.jpg",
-      }
+      },
     },
   ],
   [QuestionLayout.Descriptions]: [
@@ -77,14 +78,14 @@ const options: { [key in QuestionLayout]: Option[] } = {
       data: {
         text: "Pizza",
         description: "This is a pizza",
-      }
+      },
     },
     {
       id: "celery_id",
       data: {
         text: "Celery",
         description: "This is celery",
-      }
+      },
     },
   ],
 };
